chore: Migrate typegpu to build using tsdown
