+++ package.json
@@ -46,7 +46,6 @@
     "jiti": "catalog:build",
     "oxlint": "^1.43.0",
     "pkg-pr-new": "^0.0.62",
-    "tsup": "catalog:build",
     "typescript": "catalog:types",
     "unplugin-typegpu": "workspace:*",
     "vite-imagetools": "catalog:frontend",
