--- packages/tgpu-wgsl-parser/deno.json
@@ -1,7 +1,0 @@
-{
-  "exclude": ["."],
-  "fmt": {
-    "exclude": ["!."],
-    "singleQuote": true
-  }
-}
