--- packages/tgpu-wgsl-parser/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "tgpu-wgsl-parser",
-  "private": true,
-  "version": "0.0.0",
-  "description": "A WGSL code parser.",
-  "license": "MIT",
-  "type": "module",
-  "exports": {
-    ".": "./src/index.ts",
-    "./nearley-redirect-plugin": "./nearley-redirect-plugin.mjs"
-  },
-  "publishConfig": {
-    "directory": "dist",
-    "linkDirectory": false,
-    "main": "./dist/index.js",
-    "types": "./index.d.ts",
-    "exports": {
-      ".": {
-        "types": "./dist/index.d.ts",
-        "module": "./dist/index.js",
-        "import": "./dist/index.js",
-        "default": "./dist/index.cjs"
-      }
-    }
-  },
-  "files": ["dist"],
-  "sideEffects": false,
-  "scripts": {
-    "build": "tsup",
-    "test:types": "pnpm tsc --p ./tsconfig.json --noEmit",
-    "prepublishOnly": "tgpu-dev-cli prepack"
-  },
-  "engines": {
-    "node": ">=12.20.0"
-  },
-  "keywords": ["webgpu", "wgpu", "wgsl", "parser", "ast"],
-  "packageManager": "pnpm@10.15.1+sha512.34e538c329b5553014ca8e8f4535997f96180a1d0f614339357449935350d924e22f8614682191264ec33d1462ac21561aff97f6bb18065351c162c7e8f6de67",
-  "dependencies": {
-    "nearley": "^2.20.1"
-  },
-  "devDependencies": {
-    "@types/moo": "^0.5.9",
-    "@types/nearley": "^2.11.5",
-    "@webgpu/types": "catalog:types",
-    "moo": "^0.5.2",
-    "tsup": "catalog:build",
-    "tsx": "^4.20.6",
-    "typescript": "catalog:types",
-    "vitest": "catalog:test",
-    "@typegpu/tgpu-dev-cli": "workspace:*"
-  }
-}
