--- packages/tgpu-wgsl-parser/src/grammar.ne.d.ts
@@ -1,2 +1,0 @@
-export * from './grammar.ts';
-export { default } from './grammar.ts';
