--- packages/tgpu-wgsl-parser/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "include": ["src/**/*", "tests/**/*"],
-  "exclude": ["node_modules", "dist"]
-}
