--- packages/tgpu-wgsl-parser/vitest.config.mts
@@ -1,6 +1,0 @@
-import { defineConfig } from 'vitest/config';
-import nearleyRedirectPlugin from './nearley-redirect-plugin.mjs';
-
-export default defineConfig({
-  plugins: [nearleyRedirectPlugin()],
-});
