+++ packages/tinyest/package.json
@@ -29,7 +29,7 @@
   },
   "sideEffects": false,
   "scripts": {
-    "build": "tsup",
+    "build": "tsdown",
     "prepublishOnly": "tgpu-dev-cli prepack"
   },
   "engines": {
@@ -46,7 +46,7 @@
   "homepage": "https://docs.swmansion.com/TypeGPU",
   "devDependencies": {
     "typescript": "catalog:types",
-    "tsup": "catalog:build",
+    "tsdown": "catalog:build",
     "@typegpu/tgpu-dev-cli": "workspace:*"
   },
   "packageManager": "pnpm@10.15.1+sha512.34e538c329b5553014ca8e8f4535997f96180a1d0f614339357449935350d924e22f8614682191264ec33d1462ac21561aff97f6bb18065351c162c7e8f6de67"
