+++ packages/tinyest/tsdown.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from 'tsdown';
+
+export default defineConfig({
+  entry: ['src/index.ts'],
+  outDir: 'dist',
+  format: 'esm',
+  dts: true,
+  platform: 'neutral',
+  target: false,
+});
