+++ pnpm-workspace.yaml
@@ -7,7 +7,7 @@ catalog:
 
 catalogs:
   build:
-    tsup: ^8.5.0
+    tsdown: ^0.20.3
     unbuild: ^3.5.0
     jiti: ^2.6.0
   types:
